CsvReaderTest done and passing.
Have decided to keep the "simplecsv" interpretation of two quotes in a row
(ignoring the RFC 4180 spec).  The interpretation of explicit escapes in the
underlying data still differs between simplecsv and opencsv.  However, a one
line code change can be made to allow either, so I plan to create a new optional
setting in the parser.