CsvParser is now interface. CsvParser -> SimpleCsvParser.  Tests all passing.
Pulled footloosejava's CsvParser in as MultiLineCsvParser, but not yet refactored
to impl the CsvParser interface and no tests/supporting code pulled in for it yet.

Note: CsvParser#parse returns String[].  This should probably be changed to List<String>,
but that requires changing all unit tests, so using String[] for now.